+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/reports/sessiondays/SessionDaysReport.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.reports.sessiondays;
 
-import com.microsoft.azure.servicebus.primitives.StringUtil;
 import org.apache.commons.collections4.CollectionUtils;
 import org.elasticsearch.common.Strings;
 import uk.gov.hmcts.ecm.common.helpers.UtilHelper;
@@ -17,7 +16,6 @@ import uk.gov.hmcts.ethos.replacement.docmosis.service.referencedata.jpaservice.
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -28,7 +26,7 @@ import java.util.stream.Collectors;
 import static java.lang.Math.round;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.HEARING_STATUS_HEARD;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.OLD_DATE_TIME_PATTERN;
-import static uk.gov.hmcts.ethos.replacement.docmosis.reports.memberdays.MemberDaysReport.OLD_DATE_TIME_PATTERN3;
+import static uk.gov.hmcts.ethos.replacement.docmosis.reports.ReportCommonMethods.getHearingDuration;
 
 public class SessionDaysReport {
 
@@ -277,7 +275,7 @@ public class SessionDaysReport {
                         reportDetail.setHearingSitAlone("Sit Alone".equals(
                                 hearingTypeItem.getValue().getHearingSitAlone()) ? "Y" : "");
                         setTelCon(hearingTypeItem, reportDetail);
-                        String duration = calculateDuration(dateListedTypeItem);
+                        String duration = getHearingDuration(dateListedTypeItem);
                         reportDetail.setHearingDuration(duration);
                         reportDetail.setSessionType(getSessionType(Long.parseLong(duration)));
                         reportDetail.setHearingClerk(dateListedTypeItem.getValue().getHearingClerk());
@@ -323,40 +321,6 @@ public class SessionDaysReport {
                 && hearingTypeItem.getValue().getHearingFormat().contains("Telephone") ? "Y" : "";
         reportDetail.setHearingTelConf(telConf);
 
-    }
-
-    private LocalDateTime convertHearingTime(String dateToConvert) {
-        return dateToConvert.endsWith(".000")
-                ? LocalDateTime.parse(dateToConvert, OLD_DATE_TIME_PATTERN)
-                : LocalDateTime.parse(dateToConvert, OLD_DATE_TIME_PATTERN3);
-    }
-
-    private String calculateDuration(DateListedTypeItem c) {
-        var dateListedType = c.getValue();
-        long duration = 0;
-        long breakDuration = 0;
-
-        var hearingTimingBreak = dateListedType.getHearingTimingBreak();
-        var hearingTimingResume = dateListedType.getHearingTimingResume();
-        //If there was a break and resumption during the hearing
-        if (!StringUtil.isNullOrEmpty(hearingTimingBreak)
-                && !StringUtil.isNullOrEmpty(hearingTimingResume)) {
-            var hearingBreak = convertHearingTime(hearingTimingBreak);
-            var hearingResume = convertHearingTime(hearingTimingResume);
-            breakDuration = ChronoUnit.MINUTES.between(hearingBreak, hearingResume);
-        }
-
-        var hearingTimingStart = dateListedType.getHearingTimingStart();
-        var hearingTimingFinish = dateListedType.getHearingTimingFinish();
-        if (!StringUtil.isNullOrEmpty(hearingTimingStart)
-                && !StringUtil.isNullOrEmpty(hearingTimingFinish)) {
-            var hearingStartTime = convertHearingTime(hearingTimingStart);
-            var hearingEndTime = convertHearingTime(hearingTimingFinish);
-            long startToEndDiffInMinutes = ChronoUnit.MINUTES.between(hearingStartTime, hearingEndTime);
-            duration = startToEndDiffInMinutes - breakDuration;
-        }
-
-        return String.valueOf(duration);
     }
 
     private String getSessionType(long duration) {
